# swagger.yaml
openapi: 3.0.0
info:
  title: E-Commerce Supplements API
  version: 1.0.0
  description: Complete REST API Documentation for E-Commerce Supplements Store
servers:
  - url: http://localhost:5000/api/v1
    description: Local Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /users/{id}:
    get:
      tags: [Admin - Users]
      summary: Get user details by ID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: User data
        404:
          description: User not found
    put:
      tags: [Admin - Users]
      summary: Update user data or admin role
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Updated Name"
                email:
                  type: string
                  example: "updated@example.com"
                isAdmin:
                  type: boolean
                  example: true
      responses:
        200:
          description: User updated successfully
        404:
          description: User not found
    delete:
      tags: [Admin - Users]
      summary: Delete a user by ID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: User deleted successfully
        404:
          description: User not found

  # ================= CATEGORIES =================
...